+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 Unreleased
 **********
 * Constrained RA-sweep grouping for GSM survey SBDs
+* Adds `POST /engineering/ebs/` with to allow ExecutionBlocks to be created with some initial data, such as labels
+* Adds `PATCH /engineering/ebs/{eb_id}/labels` to allow labels to be added to ExecutionBlocks after creation
 
 14.6.0
 **********